split summary into price and summary
